Moving some tests around
